ref-filter: get rid of show_ref_array_item
When we use `git for-each-ref`, every ref will call
`show_ref_array_item()` and allocate its own final strbuf.
But we can reuse the final strbuf for each step ref's output.
Since `show_ref_array_item()` is not used in many places,
we can directly delete `show_ref_array_item()` and use the
same logic code to replace it. In this way, the caller can
clearly see how the loop work.

The performance for `git for-each-ref` on the Git repository
itself with performance testing tool `hyperfine` changes from
23.7 ms ± 0.9 ms to 22.2 ms ± 1.0 ms.

At the same time, we apply this optimization to `git tag -l`
and `git branch -l`, the `git branch -l` performance upgrade
from 5.8 ms ± 0.8 ms to 2.7 ms ± 0.2 ms and `git tag -l`
performance upgrade from 5.9 ms ± 0.4 ms to 5.4 ms ± 0.4 ms.
Since the number of tags in git.git is much more than branches,
so this shows that the optimization will be more obvious in
those repositories that contain a small number of objects.

This approach is similar to the one used by 79ed0a5
(cat-file: use a single strbuf for all output, 2018-08-14)
to speed up the cat-file builtin.

Signed-off-by: ZheNing Hu <adlternative@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>